pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datawarp"
version = "2.0.0"
description = "Deterministic NHS data ingestion"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0.0",
    "xlrd>=2.0.0",
]

[project.scripts]
datawarp = "datawarp.cli.commands:app"

[tool.setuptools.packages.find]
where = ["src"]
